review of OpenClose/Modal => ShowOnClick OpenModal
