--- colab-webapp/src/main/node/app/src/components/common/layout/OpenCloseModal.tsx
@@ -1,141 +1,0 @@
-/*
- * The coLAB project
- * Copyright (C) 2021-2023 AlbaSim, MEI, HEIG-VD, HES-SO
- *
- * Licensed under the MIT License
- */
-
-import { css, cx } from '@emotion/css';
-import * as React from 'react';
-import { NavLink, Route, Routes, useLocation, useNavigate } from 'react-router-dom';
-import { linkStyle, space_lg } from '../../../styling/style';
-import Modal from './Modal';
-import OpenClose from './OpenClose';
-
-interface OpenCloseModalProps {
-  title: React.ReactNode;
-  className?: string;
-  collapsedChildren: React.ReactNode;
-  children: (collapse: () => void) => React.ReactNode;
-  footer?: (collapse: () => void) => React.ReactNode;
-  showCloseButton?: boolean;
-  route?: string;
-  status?: 'COLLAPSED' | 'EXPANDED';
-  modalClassName?: string;
-  modalBodyClassName?: string;
-  widthMax?: boolean;
-  heightMax?: boolean;
-  onPressEnterKey?: (collapse: () => void) => void;
-}
-
-export const defaultIconClassName = css({
-  width: '28px',
-  display: 'inline-block',
-  textAlign: 'center',
-});
-
-export const modalPadding = space_lg;
-
-export default function OpenCloseModal({
-  className,
-  collapsedChildren,
-  title,
-  children,
-  footer,
-  route,
-  showCloseButton = false,
-  status = 'COLLAPSED',
-  modalClassName,
-  modalBodyClassName,
-  widthMax,
-  heightMax,
-  onPressEnterKey,
-}: OpenCloseModalProps): JSX.Element {
-  const navigate = useNavigate();
-  const location = useLocation();
-  //const match = useRouteMatch();
-
-  const onClose = React.useCallback(() => {
-    if (route != null) {
-      navigate(location.pathname.replace(new RegExp(route + '$'), ''));
-    }
-  }, [navigate, location, route]);
-
-  /* const handleEnter = (event: KeyboardEvent) => {
-    if (event.key === 'Enter') {
-      if(onPressEnterKey){
-        onPressEnterKey(() => onClose());
-      }     
-     }
-  };
-
-  React.useEffect(() => {
-    document.addEventListener('keydown', handleEnter, true);
-    return () => {
-      document.removeEventListener('keydown', handleEnter, true);
-    };
-  }); */
-
-  if (route != null) {
-    return (
-      <Routes>
-        <Route
-          path={route}
-          element={
-            <>
-              {collapsedChildren}
-              <Modal title={title} showCloseButton={showCloseButton} onClose={onClose}>
-                {onCloseModal => children(onCloseModal)}
-              </Modal>
-            </>
-          }
-        />
-        <Route
-          path="*"
-          element={
-            <NavLink
-              className={cx(
-                linkStyle,
-                css({
-                  textDecoration: 'none',
-                  '&:hover': {
-                    textDecoration: 'none',
-                  },
-                }),
-              )}
-              to={route}
-            >
-              {collapsedChildren}
-            </NavLink>
-          }
-        />
-      </Routes>
-    );
-  } else {
-    return (
-      <OpenClose
-        status={status}
-        showCollapsedChildrenWhenOpened
-        collapsedChildren={collapsedChildren}
-        className={className}
-      >
-        {onClose => (
-          <Modal
-            title={title}
-            showCloseButton={showCloseButton}
-            onClose={onClose}
-            onPressEnterKey={onPressEnterKey}
-            className={cx(
-              modalClassName,
-              widthMax && css({ width: '800px' }, heightMax && css({ height: '580px' })),
-            )}
-            modalBodyClassName={modalBodyClassName}
-            footer={footer}
-          >
-            {onCloseModal => children(onCloseModal)}
-          </Modal>
-        )}
-      </OpenClose>
-    );
-  }
-}
