+++ colab-webapp/src/main/node/app/src/components/projects/creation/ProjectCreator.tsx
@@ -14,7 +14,7 @@ import { useAppDispatch, useLoadingState } from '../../../store/hooks';
 import { space_lg, space_sm } from '../../../styling/style';
 import Button from '../../common/element/Button';
 import Flex from '../../common/layout/Flex';
-import OpenCloseModal from '../../common/layout/OpenCloseModal';
+import OpenModalOnClick from '../../common/layout/OpenModalOnClick';
 import ProjectModelSelector from '../models/ProjectModelSelector';
 import { defaultProjectIllustration } from '../ProjectCommon';
 import ProjectDataInitialization from './ProjectDataInitialization';
@@ -116,10 +116,9 @@ export default function ProjectCreator({
   }, [readOnly, status]);
 
   return (
-    <OpenCloseModal
+    <OpenModalOnClick
       title={title}
-      widthMax
-      heightMax
+      size="full"
       collapsedChildren={
         <Button kind="outline" size="sm" className={collapsedButtonClassName} icon="add">
           {i18n.modules.project.actions.createProject}
@@ -246,6 +245,6 @@ export default function ProjectCreator({
             </Flex> */}
         </>
       )}
-    </OpenCloseModal>
+    </OpenModalOnClick>
   );
 }
